Move merge functions into merge
