+++ tests/test_merge_tier2_metadata.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import numpy as np
 import pytest
-from merge_tier2_metadata import (
+from helper_files.merge import (
     tab_is_protocol,
     rename_tier2_columns,
     flatten_tier2_spreadsheet,
